Move more directors under skills
May think about retiring the concept of directors
