+++ glyph-bot/src/main/kotlin/Glyph.kt
@@ -31,17 +31,16 @@ import net.dv8tion.jda.api.sharding.DefaultShardManagerBuilder
 import net.dv8tion.jda.api.utils.cache.CacheFlag
 import org.yttr.glyph.bot.ai.AIAgent
 import org.yttr.glyph.bot.ai.dialogflow.Dialogflow
-import org.yttr.glyph.bot.directors.AuditingDirector
-import org.yttr.glyph.bot.directors.StatusDirector
-import org.yttr.glyph.bot.directors.config.ConfigDirector
-import org.yttr.glyph.bot.directors.servers.BotList
-import org.yttr.glyph.bot.directors.servers.ServerDirector
-import org.yttr.glyph.bot.directors.starboard.StarboardDirector
 import org.yttr.glyph.bot.messaging.MessagingDirector
 import org.yttr.glyph.bot.messaging.quickview.QuickviewDirector
-import org.yttr.glyph.bot.skills.FallbackSkill
+import org.yttr.glyph.bot.presentation.BotList
+import org.yttr.glyph.bot.presentation.ServerDirector
+import org.yttr.glyph.bot.presentation.StatusDirector
 import org.yttr.glyph.bot.skills.SkillDirector
+import org.yttr.glyph.bot.skills.config.ConfigDirector
+import org.yttr.glyph.bot.skills.config.ServerConfigSkill
 import org.yttr.glyph.bot.skills.creator.ChangeStatusSkill
+import org.yttr.glyph.bot.skills.moderation.AuditingDirector
 import org.yttr.glyph.bot.skills.moderation.BanSkill
 import org.yttr.glyph.bot.skills.moderation.GuildInfoSkill
 import org.yttr.glyph.bot.skills.moderation.KickSkill
@@ -54,9 +53,10 @@ import org.yttr.glyph.bot.skills.play.RedditSkill
 import org.yttr.glyph.bot.skills.roles.RoleListSkill
 import org.yttr.glyph.bot.skills.roles.RoleSetSkill
 import org.yttr.glyph.bot.skills.roles.RoleUnsetSkill
+import org.yttr.glyph.bot.skills.starboard.StarboardDirector
+import org.yttr.glyph.bot.skills.util.FallbackSkill
 import org.yttr.glyph.bot.skills.util.FeedbackSkill
 import org.yttr.glyph.bot.skills.util.HelpSkill
-import org.yttr.glyph.bot.skills.util.ServerConfigSkill
 import org.yttr.glyph.bot.skills.util.SnowstampSkill
 import org.yttr.glyph.bot.skills.util.SourceSkill
 import org.yttr.glyph.bot.skills.util.StatusSkill
