+++ glyph-bot/src/main/kotlin/presentation/ServerDirector.kt
@@ -4,7 +4,7 @@
  * Glyph, a Discord bot that uses natural language instead of commands
  * powered by DialogFlow and Kotlin
  *
- * Copyright (C) 2017-2020 by Ian Moore
+ * Copyright (C) 2017-2021 by Ian Moore
  *
  * This file is part of Glyph.
  *
@@ -22,7 +22,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package org.yttr.glyph.bot.directors.servers
+package org.yttr.glyph.bot.presentation
 
 import io.ktor.client.HttpClient
 import io.ktor.client.features.ResponseException
