+++ glyph-bot/src/main/kotlin/skills/config/ConfigDirector.kt
@@ -4,7 +4,7 @@
  * Glyph, a Discord bot that uses natural language instead of commands
  * powered by DialogFlow and Kotlin
  *
- * Copyright (C) 2017-2020 by Ian Moore
+ * Copyright (C) 2017-2021 by Ian Moore
  *
  * This file is part of Glyph.
  *
@@ -22,7 +22,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package org.yttr.glyph.bot.directors.config
+package org.yttr.glyph.bot.skills.config
 
 import com.google.gson.GsonBuilder
 import com.google.gson.JsonArray
