+++ glyph-bot/src/main/kotlin/skills/config/ServerConfigSkill.kt
@@ -1,10 +1,10 @@
 /*
- * BotList.kt
+ * ServerConfigSkill.kt
  *
  * Glyph, a Discord bot that uses natural language instead of commands
  * powered by DialogFlow and Kotlin
  *
- * Copyright (C) 2017-2020 by Ian Moore
+ * Copyright (C) 2017-2021 by Ian Moore
  *
  * This file is part of Glyph.
  *
@@ -22,22 +22,24 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package org.yttr.glyph.bot.directors.servers
+package org.yttr.glyph.bot.skills.config
+
+import net.dv8tion.jda.api.Permission
+import net.dv8tion.jda.api.events.message.MessageReceivedEvent
+import org.yttr.glyph.bot.ai.AIResponse
+import org.yttr.glyph.bot.messaging.Response
+import org.yttr.glyph.bot.skills.Skill
 
 /**
- * A bot list website's API for tracking statistics
+ * Tells people to use the config website to edit their config
  */
-data class BotList(
-    /**
-     * The name of the bot list website
-     */
-    val name: String,
-    /**
-     * The API endpoint for the bot list website
-     */
-    val apiEndpoint: String,
-    /**
-     * The token for authenticating the request to the bot list website
-     */
-    val token: String
-)
+class ServerConfigSkill : Skill(
+    "skill.config.server",
+    cooldownTime = 15,
+    guildOnly = true,
+    requiredPermissionsSelf = listOf(Permission.MANAGE_WEBHOOKS),
+    requiredPermissionsUser = listOf(Permission.ADMINISTRATOR)
+) {
+    override suspend fun onTrigger(event: MessageReceivedEvent, ai: AIResponse): Response =
+        Response.Volatile("To edit your config, visit https://gl.yttr.org/config")
+}
