+++ glyph-bot/src/main/kotlin/skills/moderation/AuditingDirector.kt
@@ -22,7 +22,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package org.yttr.glyph.bot.directors
+package org.yttr.glyph.bot.skills.moderation
 
 import net.dv8tion.jda.api.EmbedBuilder
 import net.dv8tion.jda.api.entities.Guild
