+++ glyph-bot/src/main/kotlin/skills/starboard/StarredMessage.kt
@@ -22,7 +22,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package org.yttr.glyph.bot.directors.starboard
+package org.yttr.glyph.bot.skills.starboard
 
 import club.minnced.discord.webhook.exception.HttpException
 import kotlinx.coroutines.future.await
@@ -37,10 +37,10 @@ import net.dv8tion.jda.api.events.message.guild.GenericGuildMessageEvent
 import net.dv8tion.jda.api.exceptions.PermissionException
 import org.apache.commons.lang3.StringUtils
 import org.slf4j.LoggerFactory
-import org.yttr.glyph.bot.directors.starboard.StarboardDirector.Companion.TRACKING_PREFIX
-import org.yttr.glyph.bot.directors.starboard.StarboardDirector.Companion.emojiAlias
 import org.yttr.glyph.bot.extensions.asPlainMention
 import org.yttr.glyph.bot.messaging.WebhookDirector
+import org.yttr.glyph.bot.skills.starboard.StarboardDirector.Companion.TRACKING_PREFIX
+import org.yttr.glyph.bot.skills.starboard.StarboardDirector.Companion.emojiAlias
 import org.yttr.glyph.shared.config.server.StarboardConfig
 import org.yttr.glyph.shared.pubsub.redis.RedisAsync
 import org.yttr.glyph.shared.redis.redlockLock
