+++ glyph-bot/src/main/kotlin/skills/util/FallbackSkill.kt
@@ -22,12 +22,13 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package org.yttr.glyph.bot.skills
+package org.yttr.glyph.bot.skills.util
 
 import net.dv8tion.jda.api.events.message.MessageReceivedEvent
 import net.dv8tion.jda.api.exceptions.InsufficientPermissionException
 import org.yttr.glyph.bot.ai.AIResponse
 import org.yttr.glyph.bot.messaging.Response
+import org.yttr.glyph.bot.skills.Skill
 
 /**
  * A skill that handles when a skill can't be found and there's no response from DialogFlow
